Deleting&moving files, modifying paths as part of refactor
