+++ plugin-admin/old_plugins_def.js
