--- plugin-admin/refactoring/testing.js
@@ -1,3 +1,0 @@
-const { flatPlugins } = require('./localPluginData')
-
-console.log(Object.keys(flatPlugins).length)
